refactor: Move back to `u32` for `VersionType` and everything row-related

+++ crates/concepts/src/lib.rs
@@ -2061,7 +2061,7 @@ pub struct PackageIfcFns {
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ComponentRetryConfig {
-    pub max_retries: Option<u16>,
+    pub max_retries: Option<u32>,
     pub retry_exp_backoff: Duration,
 }
 impl ComponentRetryConfig {
