+++ crates/wasm-workers/src/workflow/caching_db_connection.rs
@@ -269,7 +269,7 @@ impl CachingDbConnection {
     pub(crate) async fn subscribe_to_next_responses(
         &self,
         execution_id: &ExecutionId,
-        start_idx: u16,
+        start_idx: u32,
         timeout_fut: Pin<Box<dyn Future<Output = ()> + Send>>,
     ) -> Result<Vec<JoinSetResponseEventOuter>, DbErrorReadWithTimeout> {
         self.db_connection
